Add logs for Token security and extract path to property
Add logs to indicate when Token security is active and when not. Also
fix typo in log message and set default configuration to Certificate
security.

Extract path of JSON file from hardcoded string to a property in
application properties file. This path is referenced from the target
directory so execution of jar file from other directories will fail
sometime during runtime, when service is consumed.